+++ test/mocks/SubscriptionMock.sol
@@ -43,7 +43,7 @@ contract SubscriptionMock is Subscribable, SubscriptionCreation, ERC721Mock {
     function tip(uint256 tokenId, uint256 amount, string calldata message) external {}
 
     /// @notice "Mints" a new subscription token
-    function mint(uint256 amount, uint256 multiplier, string calldata message) external returns (uint256) {}
+    function mint(uint256 amount, uint24 multiplier, string calldata message) external returns (uint256) {}
 
     /// @notice "Burns" a subscription token, deletes all achieved subscription
     ///         data and does not withdraw any withdrawable funds
